+++ packages/@expo/cli/src/export/__tests__/exportHermes.test.ts
@@ -1,5 +1,5 @@
 import type { ExpoConfig } from '@expo/config';
-import path from 'path';
+import * as path from 'node:path';
 
 import {
   getHermesBytecodeBundleVersionAsync,
