+++ packages/@expo/cli/src/start/server/middleware/inspector/JsInspector.ts
