+++ packages/@expo/cli/src/start/server/middleware/inspector/LaunchBrowser.types.ts
