+++ packages/@expo/cli/src/start/server/middleware/inspector/__tests__/fixtures/metroInspectorResponse.ts
